split actions but keep es deps in common for api tests
