+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/services/ElasticsearchServiceTest.scala
@@ -13,6 +13,9 @@ class ElasticsearchServiceTest
     with ScalaFutures
     with WorksUtil {
 
+  val indexName = "works"
+  val itemType = "work"
+
   val searchService =
     new ElasticSearchService(indexName, itemType, elasticClient)
