+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/services/WorksServiceTest.scala
@@ -21,6 +21,9 @@ class WorksServiceTest
     with ScalaFutures
     with WorksUtil {
 
+  val indexName = "works"
+  val itemType = "work"
+
   val searchService =
     new ElasticSearchService(indexName, itemType, elasticClient)
