+++ common/docker-compose.yml
@@ -6,3 +6,12 @@ sqs:
   image: s12v/elasticmq
   ports:
     - "9324:9324"
+elasticsearch:
+  image: docker.elastic.co/elasticsearch/elasticsearch:5.4.0
+  ports:
+    - "9200:9200"
+    - "9300:9300"
+  environment:
+    - "http.host=0.0.0.0"
+    - "transport.host=0.0.0.0"
+    - "cluster.name=wellcome"
