+++ common/src/test/scala/uk/ac/wellcome/elasticsearch/WorksIndexTest.scala
@@ -1,14 +1,9 @@
-package uk.ac.wellcome.platform.ingestor.models
+package uk.ac.wellcome.elasticsearch
 
-import com.sksamuel.elastic4s.http.ElasticDsl._
-import com.sksamuel.elastic4s.mappings.dynamictemplate.DynamicMapping
-import org.elasticsearch.client.ResponseException
 import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.{BeforeAndAfterEach, FunSpec, Matchers}
-import uk.ac.wellcome.models.{IdentifierSchemes, _}
-import uk.ac.wellcome.platform.ingestor.test.utils.ElasticSearchLocal
+import uk.ac.wellcome.test.utils.ElasticSearchLocal
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
-import uk.ac.wellcome.utils.JsonUtil
 
 class WorksIndexTest
     extends FunSpec
