+++ common/src/test/scala/uk/ac/wellcome/test/utils/IndexedElasticSearchLocal.scala
@@ -1,14 +1,12 @@
-package uk.ac.wellcome.platform.ingestor.test.utils
+package uk.ac.wellcome.test.utils
 
-import com.sksamuel.elastic4s.http.ElasticDsl.{indexExists, indexInto, search}
+import com.sksamuel.elastic4s.http.ElasticDsl.{indexExists, indexInto, search, _}
 import com.sksamuel.elastic4s.http.index.IndexResponse
 import org.scalatest.{BeforeAndAfterEach, Suite}
+import uk.ac.wellcome.elasticsearch.WorksIndex
 import uk.ac.wellcome.models.Work
-import uk.ac.wellcome.platform.ingestor.models.WorksIndex
-import uk.ac.wellcome.utils.JsonUtil
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
-
-import com.sksamuel.elastic4s.http.ElasticDsl._
+import uk.ac.wellcome.utils.JsonUtil
 
 import scala.concurrent.Future
